Simple/requirements.txt
cognite-sdk
python-dotenv